requirements.txt
pymongo
python-telegram-bot
tabulate
dataclasses
requests